refactor: move Go discovery logic from workspace.js to golang_gomodules.js
Move discoverGoWorkspaceModules to its provider file so workspace.js only
retains generic scaffolding.
