+++ src/index.js
@@ -9,8 +9,8 @@ import { getCustom } from "./tools.js";
 import { resolveBatchMetadata, resolveContinueOnError } from './batch_opts.js'
 import { discoverMavenModules } from './providers/java_maven.js'
 import { discoverGradleSubprojects } from './providers/java_gradle.js'
+import { discoverGoWorkspaceModules } from './providers/golang_gomodules.js'
 import {
-	discoverGoWorkspaceModules,
 	discoverWorkspaceCrates,
 	discoverWorkspacePackages,
 	filterManifestPathsByDiscoveryIgnore,
